some format fixes in order to get it compile for m68k, tested with i386 as well.
git-svn-id: https://svn.aros.org/svn/aros/trunk@54890 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/workbench/libs/asyncio/testasync.c
@@ -20,9 +20,9 @@ int main(void)
         for (i = 0; i < 1000; i++)
         {
             bytes = WriteLineAsync(write_fh, line);
-            printf("%d bytes written.\n", bytes);
+            printf("%d bytes written.\n", (int)bytes);
             bytes = WriteCharAsync(write_fh,'\n');
-            printf("%d byte written.\n", bytes);
+            printf("%d byte written.\n", (int)bytes);
         }
         CloseAsync(write_fh);
     }
